+++ app/Services/WorkStudio/Exceptions/UserNotFoundException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Services\WorkStudio\Exceptions;
+
+use Exception;
+
+class UserNotFoundException extends Exception
+{
+    public function __construct(string $username)
+    {
+        parent::__construct("User not found in WorkStudio: {$username}");
+    }
+}